pyproject.toml
[tool.poetry]
name = "testfittin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
dadata = "^21.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
